refactor(shared): create helper is zoom link ml-148
