+++ packages/shared/src/libs/helpers/helpers.ts
@@ -1,2 +1,3 @@
 export { configureString } from "./configure-string/configure-string.helper.js";
 export { formatDate } from "./format-date/format-date.js";
+export { isZoomLink } from "./validate-zoom-link/validate-zoom-link.js";
